split complexity of LayersFilter component
