+++ apps/fishing-map/features/workspace/shared/LayerFiltersSource.tsx
@@ -2,12 +2,10 @@ import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 
 import { DatasetTypes } from '@globalfishingwatch/api-types'
-import type { SupportedDatasetFilter } from '@globalfishingwatch/datasets-client'
 import { type UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
-import type { MultiSelectOnChange, MultiSelectOption } from '@globalfishingwatch/ui-components'
+import type { MultiSelectOnChange } from '@globalfishingwatch/ui-components'
 import { MultiSelect } from '@globalfishingwatch/ui-components'
 
-import { getCommonFiltersInDataview } from 'features/dataviews/dataviews.filters'
 import { getPlaceholderBySelections } from 'features/i18n/utils'
 import { selectIsGuestUser } from 'features/user/selectors/user.selectors'
 import { useVesselGroupsOptions } from 'features/vessel-groups/vessel-groups.hooks'
@@ -18,37 +16,9 @@ import {
   getSourcesSelectedInDataview,
 } from '../activity/activity.utils'
 
+import { cleanDataviewFiltersNotAllowed } from './LayerFilters.utils'
+
 import styles from './LayerFilters.module.css'
-
-const cleanDataviewFiltersNotAllowed = (
-  dataview: UrlDataviewInstance,
-  vesselGroups: MultiSelectOption[],
-  isGuestUser?: boolean
-) => {
-  const filters = dataview.config?.filters ? { ...dataview.config.filters } : {}
-  Object.keys(filters).forEach((k) => {
-    const key = k as SupportedDatasetFilter
-    if (filters[key]) {
-      const newFilterOptions = getCommonFiltersInDataview(dataview, key, {
-        vesselGroups,
-        isGuestUser,
-      })
-      const newFilterSelection = newFilterOptions?.filter((option) =>
-        dataview.config?.filters?.[key]?.includes(option.id)
-      )
-
-      // We have to remove the key if it is not supported by the datasets selecion
-      if (newFilterOptions.length === 0) {
-        delete filters[key]
-        // or keep only the options that every dataset have in common
-      } else if (!newFilterSelection?.length !== dataview.config?.filters?.[key]?.length) {
-        filters[key] = newFilterSelection.map(({ id }) => id)
-      }
-    }
-  })
-
-  return filters
-}
 
 type LayerFiltersSourceProps = {
   dataview: UrlDataviewInstance
